fix: remove minSdkVersion < 23 code
