+++ AnkiDroid/src/main/java/com/ichi2/compat/CompatHelper.kt
@@ -45,8 +45,7 @@ class CompatHelper private constructor() {
         sdkVersion >= Build.VERSION_CODES.S -> CompatV31()
         sdkVersion >= Build.VERSION_CODES.Q -> CompatV29()
         sdkVersion >= Build.VERSION_CODES.O -> CompatV26()
-        sdkVersion >= Build.VERSION_CODES.M -> CompatV23()
-        else -> CompatV21()
+        else -> CompatV23()
     }
 
     companion object {
