+++ AnkiDroid/src/test/java/com/ichi2/compat/Test21And26.kt
@@ -31,7 +31,7 @@ import java.io.IOException
 import kotlin.test.assertFailsWith
 
 /**
- * Allows to test with CompatV21 and V26.
+ * Allows to test with CompatV23 and V26.
  * In particular it allows to test version of the code that uses [Files] and [Path] classes.
  * And versions that must restrict themselves to [File].
  */
@@ -41,7 +41,7 @@ abstract class Test21And26 {
         @JvmStatic // required for Parameters
         @Parameterized.Parameters(name = "{1}")
         fun data(): Iterable<Array<Any>> = sequence {
-            yield(arrayOf(CompatV21(), "CompatV21"))
+            yield(arrayOf(CompatV23(), "CompatV23"))
             yield(arrayOf(CompatV26(), "CompatV26"))
         }.asIterable()
 
@@ -69,8 +69,8 @@ abstract class Test21And26 {
     /** Used in the "Test Results" Window */
     lateinit var unitTestDescription: String
 
-    val isV21: Boolean
-        get() = compat is CompatV21
+    val isV23: Boolean
+        get() = compat is CompatV23
     val isV26: Boolean
         get() = compat is CompatV26
